refactor: split types with route::types
